requirements.txt
requests
beautifulsoup4
tweepy
PyYAML
python-dateutil
